Introduce supervisors for problems
A supervisor object decides how to proceed in case of a problem (indicated
by an exception): Either ignore and continue or abort the current
operation.

The framework has been prepared to handle CannotReadException type
problems and a basic STDIO-based implementation is the current default.

A corrupt .wav file has been added for tests, and tests adjusted.