Handle collisions when saving songs
Two songs with equal artist and album may also have the same
file name (maybe previously stored in different locations).
To avoid errors when trying to save a song to a path where
such a file already exists, we append a number in brackets,
similar to what e.g. Firefox does for downloads.